refactor: 💡 refactoring text and image of cargallery component
