+++ src/pages/works/components/CarGallery/index.stories.tsx
@@ -1,7 +1,7 @@
 import { CarGallery as Component } from './index';
 import { Meta, StoryObj } from '@storybook/react';
-import { swiperItemList01 } from '@/pages/works/const/swiperItemList';
 import styles from '@/pages/works/components/CarGallery/index.module.scss';
+import { swiperItemList01 } from '@/pages/works/components/CarGallery/const/carGalleryData';
 
 const meta: Meta<typeof Component> = {
   component: Component,
