+++ src/pages/works/index.page.tsx
@@ -10,7 +10,7 @@ import {
   swiperItemList01,
   swiperItemList02,
   swiperItemList03,
-} from '@/pages/works/const/swiperItemList';
+} from '@/pages/works/components/CarGallery/const/carGalleryData';
 import { ProcessTitle } from '@/pages/works/components/ProcessTitle';
 import { ProcessDescription } from '@/pages/works/components/ProcessDescription';
 import processImage01 from '@/../public/assets/works/procedure01.jpg';
@@ -18,23 +18,7 @@ import processImage02 from '@/../public/assets/works/procedure02.jpg';
 import processImage03 from '@/../public/assets/works/procedure03.jpg';
 import { SectionTitle } from '@/components/SectionTitle';
 import { Footer } from '@/components/Footer';
-
-const text = (
-  <>
-    <p>
-      NS10-H（新車限定）は、新車で納車されてから1ヶ月前後のお車が対象となるコーティングであり、ボディは磨かず、手洗い洗車後にUVAとUVBの両紫外線を軽減するガラスコーティング剤【トップコート】を施工する形になります。
-    </p>
-    <p className={styles.text}>
-      塗装の状態は非常にキレイで、所々ディーラー様が磨いたと思われるコンパウンド残りが付着していたので、その部分をキレイに取り除いて施工させていただきました。
-    </p>
-    <p className={styles.text}>
-      同時に、窓ガラス全面撥水加工も施しましたので、雨天時の視界確保に役立つと思います。窓ガラス油膜取り作業を行ってから撥水加工を施しておりますし、市販品よりも撥水力・耐久性がございますので、非常にオススメのオプションメニューとなっております。
-    </p>
-    <p className={styles.text}>
-      コーティング後は、当店からお渡しております専用メンテナンスキットを利用して正しいセオリー通りの洗車をしていただければ、長期間に渡って美しい状態を維持できると思います。
-    </p>
-  </>
-);
+import { text } from '@/pages/works/components/CarGallery/const/carGalleryData';
 
 export default function Works() {
   return (
